backend/requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
